Changes made
Commented the reason for adding 2*Step in calculating the max limit inside groupDataAndCalculateCumulativeValues function in calculate.py file
Max years is treated as maximum no. of years the analysis should be done rather than no. of times the analysis should be done
The plot style for regression analysis is changed and 'Error' is renamed as 'Residue'